requirements.txt
flask
tensorflow==2.12.0
numpy
pillow
gunicorn